refactor: move mapPreview reducer to ui slice, set color
