+++ src/components/search/searchArea/searchBox.tsx
@@ -20,9 +20,8 @@ import { makeStyles } from "@mui/styles";
 import tailwindConfig from "../../../../tailwind.config";
 import resolveConfig from "tailwindcss/resolveConfig";
 import { AppDispatch, RootState } from "@/store";
-import { setQuery, fetchSuggestions } from "@/store/slices/searchSlice";
-import { setInputValue } from "@/store/slices/searchSlice";
-import { setShowInfoPanel, setShowClearButton, setShowDetailPanel } from "@/store/slices/uiSlice";
+import { setQuery, fetchSuggestions, setInputValue } from "@/store/slices/searchSlice";
+import { setShowInfoPanel, setShowClearButton, setShowDetailPanel, setMapPreview } from "@/store/slices/uiSlice";
 import SpellCheckMessage from "./spellCheckMessage";
 interface Props {
   schema: any;
@@ -154,6 +153,7 @@ const SearchBox = ({ schema }: Props): JSX.Element => {
     dispatch(setQuery(null));
     dispatch(setShowDetailPanel(null));
     dispatch(setShowClearButton(false));
+    dispatch(setMapPreview([]))
   };
   const isIOS = React.useMemo(() => {
     if (
