+++ src/middleware/filterHelper.ts
@@ -172,6 +172,7 @@ export const resetFilters = async (store: any) => {
   filterActions.forEach((action) => {
     store.dispatch(action);
   });
+  store.dispatch({type: "search/setMapPreview", payload: []})
   if (isBrowser) {
     const searchParams = new URLSearchParams(window.location.search);
     Object.entries(actionConfig)
