+++ src/store/slices/uiSlice.ts
@@ -1,5 +1,10 @@
 import { createSlice } from "@reduxjs/toolkit";
 
+interface MapPreviewLyr {
+  lyrId: string;
+  filterIds: string[];
+}
+
 interface UIState {
   pageFirstLoad: boolean;
   showInfoPanel: boolean;
@@ -7,6 +12,7 @@ interface UIState {
   showFilter: boolean;
   showClearButton: boolean;
   showSharedLink: boolean;
+  mapPreview: MapPreviewLyr[];
 }
 
 const initialState: UIState = {
@@ -16,6 +22,7 @@ const initialState: UIState = {
   showFilter: false,
   showClearButton: false,
   showSharedLink: false,
+  mapPreview: [],
 };
 
 const uiSlice = createSlice({
@@ -40,6 +47,9 @@ const uiSlice = createSlice({
     setShowSharedLink: (state, action) => {
       state.showSharedLink = action.payload;
     },
+    setMapPreview: (state, action) => {
+      state.mapPreview = action.payload;
+    },
   },
 });
 
@@ -50,6 +60,7 @@ export const {
   setShowFilter,
   setShowClearButton,
   setShowSharedLink,
+  setMapPreview,
 } = uiSlice.actions;
 
 export default uiSlice.reducer;
