+++ src/store/types/search.ts
@@ -1,8 +1,3 @@
-interface MapPreviewLyr {
-  lyrId: string;
-  filterIds: string[];
-}
-
 export interface SearchState {
   schema: any;
   // object
@@ -30,7 +25,6 @@ export interface SearchState {
   // map
   visOverlays: string[];
   visLyrs: string[];
-  mapPreview: MapPreviewLyr[];
   bbox: [number, number, number, number] | null;
 
   // status
@@ -55,7 +49,6 @@ export const initialState: SearchState = {
   spatialResolution: [],
   visOverlays: [],
   visLyrs: [],
-  mapPreview: [],
   indexYear: [],
   spellCheck: "",
   originalQuery: "",
